trigger_wb_update_with_secondary: use keep-latest for base packages
Filter the base packages in trigger_wb_update_with_secondary() with
keep-latest. This matches the apt behaviour, which only consider the
latest version of a package in a suite.

This also prevent two packages with the same name and the same version,
but differing by the architecture (all/any) to break dose-debbuildcheck.
While this case is probably not valid, it technically happens.